Test the code using a "teststatement"
The test statement will also serve as a blueprint for how I am to actually structure this thing. I am thinking that each section of the story needs to be based on a template, with the "story elements" randomly selected based on traits. 

I also need to have an array holding ALL of a character's traits, and sample that, to increase the number of possible variations. Because seeing the same templates will get boring fast, unless I switch up the words a bit.