Refactor the "total array"
Was inspired by pilcrow's answer in http://stackoverflow.com/questions/1801516/how-do-you-add-an-array-to-another-array-in-ruby-and-not-end-up-with-a-multi-dim. The method I previously used works just fine in creating an array that holds all possible character traits, but it was convoluted and did not make a whole lot of sense

I also attached the "super" command to all the character classes in this file, so that every character will be able to have a "total" array.