Add some adjectives from Rantionary
Rantionary is a dictionary that shows all words that are used in the random text generator "Rant". It is licensed under the MIT, and I may be able to use it for any purpose. These words are good at generating filler for a novel. I will need to test to see if I have added these words in correctly, as well as learn how to actually scrape content (right now, all I'm doing is copying each word and pasting it into the text editor..an incredibly inefficent process, but at least it works.)