Add a bunch of placeholder words over to the SelfVerb class
Now we can generate a sensible story, so long as the characters do not try to target each other
(as we did not place placeholder verbs for targeting). However, the number of Lines in this code
showed me that I may have to refactor this code greatly in order for expansion to work. I may
need to stuff all this data into some easily extendable database and retreive from this database
as approriate. This would mean mastering SQL and Rails, languages that I have not yet learned.

So, until I learn SQL and Rails, I will simply focus on the much easier task of testing this code
and coding in a plot generation system!